+++ custom_components/solarcharger/entity.py
@@ -10,18 +10,18 @@ from homeassistant.util import slugify
 
 from .config_utils import get_device_domain, is_api_defined_solarcharger_entity
 from .const import (
-    CHARGER_DOMAIN_ESPHOME,
-    CHARGER_DOMAIN_MQTT,
-    CHARGER_DOMAIN_OCPP,
-    CHARGER_DOMAIN_TESLA_CUSTOM,
-    CHARGER_DOMAIN_TESLA_FLEET,
-    CHARGER_DOMAIN_TESLA_TESSIE,
     CONFIG_NAME_GLOBAL_DEFAULTS,
     CONFIG_URL,
     DEVICE_MODEL_MAP,
     DOMAIN,
+    DOMAIN_OCPP,
+    DOMAIN_TESLA_CUSTOM,
+    DOMAIN_TESLA_FLEET,
+    DOMAIN_TESLA_TESSIE,
     ICON,
     MANUFACTURER,
+    SUBDOMAIN_ESPHOME_TESLA_BLE,
+    SUBDOMAIN_MQTT_TESLA_BLE,
     SUBENTRY_CHARGER_TYPES,
     VERSION,
 )
@@ -64,12 +64,12 @@ class SolarChargerEntityType(Enum):
 
     # For specify device
     # Only the following can be placed in a list when defining entities. All others must be single entity type.
-    TYPE_LOCAL_OCPP = CHARGER_DOMAIN_OCPP
-    TYPE_LOCAL_TESLA_CUSTOM = CHARGER_DOMAIN_TESLA_CUSTOM
-    TYPE_LOCAL_TESLA_MQTTBLE = CHARGER_DOMAIN_MQTT
-    TYPE_LOCAL_TESLA_ESPBLE = CHARGER_DOMAIN_ESPHOME
-    TYPE_LOCAL_TESLA_FLEET = CHARGER_DOMAIN_TESLA_FLEET
-    TYPE_LOCAL_TESLA_TESSIE = CHARGER_DOMAIN_TESLA_TESSIE
+    TYPE_LOCAL_OCPP = DOMAIN_OCPP
+    TYPE_LOCAL_TESLA_CUSTOM = DOMAIN_TESLA_CUSTOM
+    TYPE_LOCAL_TESLA_MQTTBLE = SUBDOMAIN_MQTT_TESLA_BLE
+    TYPE_LOCAL_TESLA_ESPBLE = SUBDOMAIN_ESPHOME_TESLA_BLE
+    TYPE_LOCAL_TESLA_FLEET = DOMAIN_TESLA_FLEET
+    TYPE_LOCAL_TESLA_TESSIE = DOMAIN_TESLA_TESSIE
     TYPE_LOCAL_USER_CUSTOM = DOMAIN
 
     #####################################
@@ -252,6 +252,8 @@ class SolarChargerEntity(Entity):
         device_model = DEVICE_MODEL_MAP.get(domain)
 
         self._attr_device_info = DeviceInfo(
+            # Best to use a non-changing variable as identifier in the device registry.
+            # identifiers={(DOMAIN, self._subentry.unique_id)},
             identifiers={(DOMAIN, self._subentry.subentry_id)},
             name=self._subentry.title,
             manufacturer=MANUFACTURER,